Rename client.py -> render_client.py
