+++ src/python/janelia_emrp/msem/wafer_60_gc_upload/details/__init__.py
@@ -1,6 +1,6 @@
 from .image_io import load_images_as_stack, store_beam_shading
 from .gc_writer import MsemCloudWriter
-from .client import MsemClient
+from .render_client import MsemClient
 from .config import AcquisitionConfig, BeamConfig, Slab
 
 __all__ = ['MsemCloudWriter', 'load_images_as_stack', 'store_beam_shading',
